fix(agent): update execution test imports after rename
